Initialization methods for Sweep class.
